food-ration/adapter/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pymysql==1.1.0
pika==1.3.2
python-dotenv==1.0.0
SQLAlchemy==2.0.23
cryptography==41.0.4
python-jose==3.3.0
schedule==1.2.0
Jinja2==3.1.2
aiofiles==23.2.1
uvicorn
flask==2.3.3